pyproject.toml
[project]
name = "netcollector"
version = "0.1.0"
description = "Collect state of network devices"
readme = "README.md"
authors = [
    { name = "Ryan Merolle"}
]
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.0.0",
    "asyncssh>=2.21.0",
    "devtools",
    "duckdb>=1.3.0",
    "importlib-resources>=6.5.2",
    "netutils>=1.13.0",
    "ntc-templates>=7.9.0",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pydantic-settings>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "scrapli-community>=2025.1.30",
    "scrapli>=2025.1.30",
    "textfsm",
    "typer>=0.16.0",
]

[project.scripts]
netcollector = "netcollector.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "coverage>=7.8.2",
    "pytest-mock>=3.14.0",
    "pytest>=8.3.5",
    "ruff>=0.11.11",
    "ty>=v0.0.1a7",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["src"]
